+++ numerapi/signalsapi.py
@@ -2,7 +2,6 @@
 
 from typing import List, Dict, Tuple, Union
 import os
-import decimal
 from io import BytesIO
 
 import requests
@@ -151,49 +150,6 @@ class SignalsAPI(base_api.Api):
         create = self.raw_query(create_query, arguments, authorization=True)
         return create['data']['createSignalsSubmission']['id']
 
-    def public_user_profile(self, username: str) -> Dict:
-        """Fetch the public Numerai Signals profile of a user.
-
-        Args:
-            username (str)
-
-        Returns:
-            dict: user profile including the following fields:
-
-                * username (`str`)
-                * startDate (`datetime`)
-                * id (`string`)
-                * bio (`str`)
-                * nmrStaked (`decimal.Decimal`)
-
-        Example:
-            >>> api = SignalsAPI()
-            >>> api.public_user_profile("floury_kerril_moodle")
-            {'bio': None,
-             'id': '635db2a4-bdc6-4e5d-b515-f5120392c8c9',
-             'startDate': datetime.datetime(2019, 3, 26, 0, 43),
-             'username': 'floury_kerril_moodle',
-             'nmrStaked': Decimal('14.630994874320760131')}
-
-        """
-        query = """
-          query($username: String!) {
-            v2SignalsProfile(modelName: $username) {
-              id
-              startDate
-              username
-              bio
-              nmrStaked
-            }
-          }
-        """
-        arguments = {'username': username}
-        data = self.raw_query(query, arguments)['data']['v2SignalsProfile']
-        # convert strings to python objects
-        utils.replace(data, "startDate", utils.parse_datetime_string)
-        utils.replace(data, "nmrStaked", utils.parse_float_string)
-        return data
-
     def daily_model_performances(self, username: str) -> List[Dict]:
         """Fetch daily Numerai Signals performance of a model.
 
@@ -283,19 +239,3 @@ class SignalsAPI(base_api.Api):
         """
         path = self.download_dataset("signals/v1.0/live.parquet")
         return pd.read_parquet(path).numerai_ticker.tolist()
-
-    def stake_get(self, username) -> decimal.Decimal:
-        """get current stake for a given users
-
-        Args:
-            username (str)
-
-        Returns:
-            decimal.Decimal: current stake
-
-        Example:
-            >>> SignalsAPI().stake_get("uuazed")
-            Decimal('14.63')
-        """
-        data = self.public_user_profile(username)
-        return data['totalStake']